login and withdraw validations

+++ server/api/v1/users/users-email-login.js
@@ -50,52 +50,59 @@ module.exports = async (req, res) => {
 			let user = await Users.findOne({ email: params.email, password: params.password }).lean();
 			
 			if (user) {
-				var finalResponse ={};
-				finalResponse = user;
+				if(user.status == 1) {
+					var finalResponse ={};
+					finalResponse = user;
+	
+					
+					let tokendata = {};
+	
+					tokendata.language	=	user.language;
+					tokendata._id		=	user._id;
+					tokendata.id		=	user._id;
+					tokendata.phone		=	user.phone;
+					tokendata.email		=	user.email;
+					await Tokens.deleteMany({"userId":ObjectId(user._id)});
+					let token = await generateClientToken(tokendata);
+					
+	
+					await Users.updateOne({ _id: user._id }, { $set: { otp: '', otp_time: '', token: token, device_id: params.device_id, device_type: params.device_type } });
+					response["message"] = "Logged-in successfully.";         
+	
+					let tokenInsertData		=	{};
+					tokenInsertData.userId	=	new ObjectId(user._id);
+					tokenInsertData.token	=	token;
+					tokenInsertData.device_id	=	params.device_id;
+					tokenInsertData.device_type	=	params.device_type;
+					
+					finalResponse.token = token;
 
-				
-				let tokendata = {};
-
-				tokendata.language	=	user.language;
-				tokendata._id		=	user._id;
-				tokendata.id		=	user._id;
-				tokendata.phone		=	user.phone;
-				tokendata.email		=	user.email;
-				await Tokens.deleteMany({"userId":ObjectId(user._id)});
-				let token = await generateClientToken(tokendata);
-				
-
-				await Users.update({ _id: user._id }, { $set: { otp: '', otp_time: '', token: token, device_id: params.device_id, device_type: params.device_type } });
-				response["message"] = "Logged-in successfully.";         
-
-				let tokenInsertData		=	{};
-				tokenInsertData.userId	=	new ObjectId(user._id);
-				tokenInsertData.token	=	token;
-				tokenInsertData.device_id	=	params.device_id;
-				tokenInsertData.device_type	=	params.device_type;
-				
-				finalResponse.token = token;
-
-                                try{
-					let referalUser = await ReferralCodeDetails.findOne({ user_id: user._id });
-					if (referalUser && referalUser.referal_code) {
-						finalResponse.refered_by_code = referalUser.referal_code;
+					try{
+						let referalUser = await ReferralCodeDetails.findOne({ user_id: user._id });
+						if (referalUser && referalUser.referal_code) {
+							finalResponse.refered_by_code = referalUser.referal_code;
+						}
+					} catch(errrrr){
+						
 					}
-				}catch(errrrr){}
-
-				//****************Set Toen In Redis**************** */
-				var newTokenObj = {user_id : user._id, token : token}
-				redis.setRedisLogin(RedisKeys.USER_AUTH_CHECK + user._id, newTokenObj);
-				//******************************* */
-
-				Tokens.create(tokenInsertData);
-				delete finalResponse.password;
-				delete finalResponse.otp; 
-				
-				response["status"]	=	true;
-				response["token"]	=	token;
-				response["data"]	=	finalResponse;
-				return res.json(response);
+	
+					//****************Set Toen In Redis**************** */
+					var newTokenObj = {user_id : user._id, token : token}
+					redis.setRedisLogin(RedisKeys.USER_AUTH_CHECK + user._id, newTokenObj);
+					//******************************* */
+	
+					Tokens.create(tokenInsertData);
+					delete finalResponse.password;
+					delete finalResponse.otp; 
+					
+					response["status"]	=	true;
+					response["token"]	=	token;
+					response["data"]	=	finalResponse;
+					return res.json(response);
+				} else {
+					response["message"] = "Please verify your phone number before email login.";
+					return res.json(response);
+				}
 
 			} else {
 				response["message"] = "Invalid login details";
